requirements.txt
torch==1.9.0+cpu
torchvision==0.10.0+cpu
numpy
pillow
